+++ run.sh
@@ -4,18 +4,18 @@
 # PE_DIM=40
 # SCENARIO=7
 # SEED=1
-# CUDA_VISIBLE_DEVICES=0 python Main.py\
-#         --data-name ml_100k\
-#         --epochs 40\
+# CUDA_VISIBLE_DEVICES=2 python Main.py\
+#         --data-name yahoo_music\
+#         --epochs 20\
 #         --save-appendix _${PE_DIM}_${SCENARIO}\
 #         --data-appendix _${PE_DIM}\
 #         --pe-dim ${PE_DIM}\
 #         --ensemble\
 #         --testing\
-#         --max-nodes-per-hop 200\
-#         --batch-size 30\
+#         --batch-size 50\
 #         --seed ${SEED}\
 #         --scenario ${SCENARIO}
+        # --max-nodes-per-hop 200\
 
 # CUDA_VISIBLE_DEVICES=2 python Main.py\
 #         --data-name ml_100k\
@@ -31,7 +31,7 @@
         # --dynamic-train\
         # --no-train\
 
-python bot.py --scenario 7
-python bot.py --scenario 8
-python bot.py --scenario 9
-python bot.py --scenario 10
+python bot.py --scenario 1 -g 2 & python bot.py --scenario 2 -g 2
+python bot.py --scenario 3 -g 2 & python bot.py --scenario 4 -g 2
+python bot.py --scenario 5 -g 2 & python bot.py --scenario 6 -g 2
+python bot.py --scenario 7 -g 2 & python bot.py --scenario 8 -g 2